Move PET related function to dedicated module
- no code changes, only re-arrangments
