+++ SW_Flow.c
@@ -120,6 +120,8 @@
 #include "SW_VegProd.h"
 #include "SW_Weather.h"
 #include "SW_Sky.h"
+
+#include "SW_Flow_lib_PET.h"
 #include "SW_Flow.h"
