+++ makefile
@@ -107,7 +107,7 @@ gtest_LDLIBS = -l$(gtest)
 sources_core = SW_Main_lib.c SW_VegEstab.c SW_Control.c generic.c \
 					rands.c Times.c mymemory.c filefuncs.c SW_Files.c SW_Model.c \
 					SW_Site.c SW_SoilWater.c SW_Markov.c SW_Weather.c SW_Sky.c \
-					SW_VegProd.c SW_Flow_lib.c SW_Flow.c SW_Carbon.c
+					SW_VegProd.c SW_Flow_lib_PET.c SW_Flow_lib.c SW_Flow.c SW_Carbon.c
 
 sources_lib = $(sw_sources) $(sources_core) SW_Output.c SW_Output_get_functions.c
 objects_lib = $(sources_lib:.c=.o)
